// client/src/components/signup-pages/interfaces/signupInterface.ts
export interface UserInfoProps{
    userid: string;
    name: string;
    phonenumber: string;
    address: string;
    password: string;
}
export interface SellerInfoProps{
    userid: string;
    name: string;
    phonenumber: string;
    business_number: string;
    account_number: string;
    password: string;
}